+++ src/aux-common/Files/FileCalculations.ts
@@ -41,7 +41,10 @@ import {
 } from './FileProxy';
 
 /// <reference path="../typings/global.d.ts" />
-import formulaLib from '../Formulas/formula-lib';
+import formulaLib, {
+    setCalculationContext,
+    getCalculationContext,
+} from '../Formulas/formula-lib';
 import SandboxInterface, { FilterFunction } from '../Formulas/SandboxInterface';
 import { PartialFile } from '../Files';
 import { FilesState, cleanFile, hasValue } from './FilesChannel';
@@ -1052,6 +1055,7 @@ export function createCalculationContextFromState(
  */
 export function createCalculationContext(
     objects: Object[],
+    userId: string = null,
     lib: SandboxLibrary = formulaLib,
     setValueHandlerFactory?: (file: File) => SetValueHandler
 ): FileCalculationContext {
@@ -1061,6 +1065,7 @@ export function createCalculationContext(
     };
     context.sandbox.interface = new SandboxInterfaceImpl(
         context,
+        userId,
         setValueHandlerFactory
     );
     return context;
@@ -1121,6 +1126,120 @@ export function filterMatchesArguments(
         }
     }
     return false;
+}
+
+/**
+ * Determines if the given username is in the username list in the given file and tag.
+ * @param calc The file calculation context.
+ * @param file The file.
+ * @param tag The tag.
+ * @param username The username to check.
+ */
+export function isInUsernameList(
+    calc: FileCalculationContext,
+    file: File,
+    tag: string,
+    username: string
+): boolean {
+    const list = getFileUsernameList(calc, file, tag);
+    return list.indexOf(username) >= 0;
+}
+
+/**
+ * Gets a list of usernames from the given file and tag.
+ * @param calc The file calculation context.
+ * @param file The file.
+ * @param tag The tag.
+ */
+export function getFileUsernameList(
+    calc: FileCalculationContext,
+    file: File,
+    tag: string
+): string[] {
+    let value = calculateFileValue(calc, file, tag);
+
+    if (value && !Array.isArray(value)) {
+        value = [value];
+    }
+
+    if (value) {
+        for (let i = 0; i < value.length; i++) {
+            let v = value[i];
+            if (isFile(v)) {
+                value[i] = v.tags['aux._user'] || v.id;
+            }
+        }
+    }
+
+    return value;
+}
+
+/**
+ * Determines if the whitelist on the given file allows the given username.
+ * Whitelists work by allowing only the usernames that are explicitly listed.
+ * If the whitelist is empty, then everything is allowed.
+ * @param calc The file calculation context.
+ * @param file The file.
+ * @param username The username to check.
+ */
+export function whitelistAllowsAccess(
+    calc: FileCalculationContext,
+    file: File,
+    username: string
+): boolean {
+    const list = getFileWhitelist(calc, file);
+    if (list) {
+        return isInUsernameList(calc, file, 'aux.whitelist', username);
+    }
+    return true;
+}
+
+/**
+ * Determines if the whitelist on the given file allows the given username.
+ * Blacklists work by denying only the usernames that are explicitly listed.
+ * If the blacklist is empty, then everything is allowed.
+ * @param calc The file calculation context.
+ * @param file The file.
+ * @param username The username to check.
+ */
+export function blacklistAllowsAccess(
+    calc: FileCalculationContext,
+    file: File,
+    username: string
+): boolean {
+    const list = getFileBlacklist(calc, file);
+    if (list) {
+        return !isInUsernameList(calc, file, 'aux.blacklist', username);
+    }
+    return true;
+}
+
+/**
+ * Gets the aux.whitelist tag from the given file.
+ * Always returns an array of strings.
+ * If any files returned by the formula, then the aux._user tag will be used from the file.
+ * @param calc The file calculation context.
+ * @param file The file.
+ */
+export function getFileWhitelist(
+    calc: FileCalculationContext,
+    file: File
+): string[] {
+    return getFileUsernameList(calc, file, 'aux.whitelist');
+}
+
+/**
+ * Gets the aux.blacklist tag from the given file.
+ * Always returns an array of strings.
+ * If any files returned by the formula, then the aux._user tag will be used from the file.
+ * @param calc The file calculation context.
+ * @param file The file.
+ */
+export function getFileBlacklist(
+    calc: FileCalculationContext,
+    file: File
+): string[] {
+    return getFileUsernameList(calc, file, 'aux.blacklist');
 }
 
 /**
@@ -1865,7 +1984,15 @@ export function isFileInContext(
     if (!contextId) return false;
 
     let result: boolean;
-    const contextValue = calculateFileValue(context, file, contextId);
+
+    let contextValue = calculateFileValue(context, file, contextId.valueOf());
+
+    if (
+        typeof contextValue === 'object' &&
+        typeof contextValue.valueOf === 'function'
+    ) {
+        contextValue = contextValue.valueOf();
+    }
 
     if (typeof contextValue === 'string') {
         result = contextValue === 'true';
@@ -2069,6 +2196,9 @@ function _calculateFormulaValue(
     formula: string,
     unwrapProxy: boolean = true
 ) {
+    const prevCalc = getCalculationContext();
+    setCalculationContext(context);
+
     const result = context.sandbox.run(
         formula,
         {
@@ -2078,6 +2208,8 @@ function _calculateFormulaValue(
         },
         convertToFormulaObject(context, object)
     );
+
+    setCalculationContext(prevCalc);
 
     if (unwrapProxy) {
         // Unwrap the proxy object
@@ -2121,6 +2253,7 @@ function _singleOrArray<T>(values: T[]) {
 }
 
 class SandboxInterfaceImpl implements SandboxInterface {
+    private _userId: string;
     objects: Object[];
     context: FileCalculationContext;
     setValueHandlerFactory: (file: File) => SetValueHandler;
@@ -2128,10 +2261,12 @@ class SandboxInterfaceImpl implements SandboxInterface {
 
     constructor(
         context: FileCalculationContext,
+        userId: string,
         setValueHandlerFactory?: (file: File) => SetValueHandler
     ) {
         this.objects = context.objects;
         this.context = context;
+        this._userId = userId;
         this.proxies = new Map();
         this.setValueHandlerFactory = setValueHandlerFactory;
     }
@@ -2185,6 +2320,10 @@ class SandboxInterfaceImpl implements SandboxInterface {
 
     uuid(): string {
         return uuid();
+    }
+
+    userId(): string {
+        return this._userId;
     }
 
     private _filterValues(values: any[], filter: FilterFunction) {