+++ src/aux-server/aux-web/aux-player/interaction/PlayerInteractionManager.ts
@@ -38,15 +38,54 @@ export class PlayerInteractionManager extends BaseInteractionManager {
         hit: Intersection
     ): IOperation {
         if (gameObject instanceof AuxFile3D) {
+            let faceValue: string = 'Unknown Face';
+
+            // Based on the normals of the file the raycast hit, determine side of the cube
+            if (hit.face.normal.x != 0) {
+                if (hit.face.normal.x > 0) {
+                    faceValue = 'left';
+                } else {
+                    faceValue = 'right';
+                }
+            } else if (hit.face.normal.y != 0) {
+                if (hit.face.normal.y > 0) {
+                    faceValue = 'top';
+                } else {
+                    faceValue = 'bottom';
+                }
+            } else if (hit.face.normal.z != 0) {
+                if (hit.face.normal.z > 0) {
+                    faceValue = 'front';
+                } else {
+                    faceValue = 'back';
+                }
+            }
+
             let fileClickOp = new PlayerFileClickOperation(
                 <PlayerSimulation3D>gameObject.contextGroup.simulation,
                 this,
-                gameObject
+                gameObject,
+                faceValue
             );
             return fileClickOp;
         } else {
             return null;
         }
+    }
+
+    handlePointerEnter(file: File, simulation: Simulation): IOperation {
+        simulation.helper.action('onPointerEnter', [file]);
+        return null;
+    }
+
+    handlePointerExit(file: File, simulation: Simulation): IOperation {
+        simulation.helper.action('onPointerExit', [file]);
+        return null;
+    }
+
+    handlePointerDown(file: File, simulation: Simulation): IOperation {
+        simulation.helper.action('onPointerDown', [file]);
+        return null;
     }
 
     createEmptyClickOperation(): IOperation {