+++ src/aux-server/aux-web/aux-projector/interaction/BuilderInteractionManager.ts
@@ -31,10 +31,14 @@ import {
     createFile,
     isContext,
     getFileConfigContexts,
+    filesInContext,
+    AuxObject,
+    toast,
+    PartialFile,
 } from '@casual-simulation/aux-common';
 import { BuilderFileClickOperation } from '../../aux-projector/interaction/ClickOperation/BuilderFileClickOperation';
 import { Physics } from '../../shared/scene/Physics';
-import { flatMap, minBy, keys } from 'lodash';
+import { flatMap, minBy, keys, uniqBy } from 'lodash';
 import {
     Axial,
     realPosToGridPos,
@@ -58,6 +62,7 @@ import MiniFile from '../MiniFile/MiniFile';
 import FileTag from '../FileTag/FileTag';
 import FileTable from '../FileTable/FileTable';
 import { appManager } from '../../shared/AppManager';
+import { Simulation } from 'aux-web/shared/Simulation';
 
 export class BuilderInteractionManager extends BaseInteractionManager {
     // This overrides the base class IGameView
@@ -233,19 +238,38 @@ export class BuilderInteractionManager extends BaseInteractionManager {
      */
     updateTileHeightAtGridPosition(
         file: ContextGroup3D,
-        position: Axial,
+        positions: Axial[],
         height: number
     ) {
-        const key = posToKey(position);
-        this._gameView.simulation3D.simulation.helper.updateFile(file.file, {
+        let partial: PartialFile = {
             tags: {
-                [`aux.context.grid`]: {
-                    [key]: {
-                        height: height,
-                    },
-                },
+                [`aux.context.grid`]: {},
             },
-        });
+        };
+
+        for (let i = 0; i < positions.length; i++) {
+            const key = posToKey(positions[i]);
+            partial.tags[`aux.context.grid`][key] = {
+                height: height,
+            };
+        }
+
+        this._gameView.simulation3D.simulation.helper.updateFile(
+            file.file,
+            partial
+        );
+    }
+
+    handlePointerEnter(file: File, simulation: Simulation): IOperation {
+        return null;
+    }
+
+    handlePointerExit(file: File, simulation: Simulation): IOperation {
+        return null;
+    }
+
+    handlePointerDown(file: File, simulation: Simulation): IOperation {
+        return null;
     }
 
     /**
@@ -408,9 +432,8 @@ export class BuilderInteractionManager extends BaseInteractionManager {
                     actions.push({
                         label: 'Raise',
                         onClick: () =>
-                            this.updateTileHeightAtGridPosition(
+                            this.SetAllHexHeight(
                                 gameObject,
-                                tile,
                                 currentHeight + increment
                             ),
                     });
@@ -418,9 +441,8 @@ export class BuilderInteractionManager extends BaseInteractionManager {
                         actions.push({
                             label: 'Lower',
                             onClick: () =>
-                                this.updateTileHeightAtGridPosition(
+                                this.SetAllHexHeight(
                                     gameObject,
-                                    tile,
                                     currentHeight - increment
                                 ),
                         });
@@ -444,6 +466,11 @@ export class BuilderInteractionManager extends BaseInteractionManager {
                 actions.push({
                     label: minimizedLabel,
                     onClick: () => this._toggleWorkspace(calc, gameObject),
+                });
+
+                actions.push({
+                    label: 'Copy',
+                    onClick: () => this._copyWorkspace(calc, gameObject),
                 });
 
                 actions.push({
@@ -474,6 +501,20 @@ export class BuilderInteractionManager extends BaseInteractionManager {
     }
 
     /**
+     * On raise or lower, set all hexes in workspace to given height
+     * @param file
+     */
+    private SetAllHexHeight(gameObject: ContextGroup3D, height: number) {
+        if (gameObject instanceof BuilderGroup3D) {
+            let tiles = gameObject.surface.hexGrid.hexes.map(
+                hex => hex.gridPosition
+            );
+
+            this.updateTileHeightAtGridPosition(gameObject, tiles, height);
+        }
+    }
+
+    /**
      * Minimizes or maximizes the given workspace.
      * @param file
      */
@@ -490,6 +531,29 @@ export class BuilderInteractionManager extends BaseInteractionManager {
                         [`aux.context.minimized`]: minimized,
                     },
                 }
+            );
+        }
+    }
+
+    /**
+     * Copies all the files on the workspace to the given user's clipboard.
+     * @param file
+     */
+    private async _copyWorkspace(
+        calc: FileCalculationContext,
+        file: ContextGroup3D
+    ) {
+        if (file && isContext(calc, file.file)) {
+            const contexts = getFileConfigContexts(calc, file.file);
+            const files = flatMap(contexts, c => filesInContext(calc, c));
+            const deduped = uniqBy(files, f => f.id);
+            await appManager.copyFilesFromSimulation(
+                file.simulation.simulation,
+                <AuxObject[]>deduped
+            );
+
+            await file.simulation.simulation.helper.transaction(
+                toast('Worksurface Copied!')
             );
         }
     }